requirements.txt
django
djangorestframework
pandas
numpy
environs
